chore(web): delete legacy api routes and backend deps — web is pure UI
- Delete apps/web/src/app/api/ directory completely
- Delete apps/web/src/app/utils/errorHandler.ts
- Remove @costwise/db, @costwise/domain, drizzle-orm, pg, bcrypt, @vercel/blob from apps/web/package.json
- Remove @costwise/db and @costwise/domain from apps/web/next.config.ts transpilePackages
- Move analytics domain types to @costwise/shared/specialTypes.ts and update web components
- Clean up redundant dotenv configs in db and api entrypoints
- Codify dev-server restart rule in docs/AGENTS.md
- Verification gates:
  - grep -rn "@costwise/domain" apps/web (0 matches outside .next cache)
  - ls apps/web/src/app/api (does not exist)
  - grep -rn "revalidatePath" apps/web (0 matches)
  - turbo run build (6/6 packages green)
  - turbo run test (5/5 suites, 126 tests green)
  - turbo run lint (green)
