+++ apps/api/src/auth.ts
@@ -1,10 +1,3 @@
-import dotenv from "dotenv";
-import path from "path";
-
-dotenv.config();
-dotenv.config({ path: path.resolve(__dirname, "../../.env") });
-dotenv.config({ path: path.resolve(__dirname, "../../../.env") });
-
 import { betterAuth } from "better-auth";
 import { drizzleAdapter } from "better-auth/adapters/drizzle";
 import bcrypt from "bcrypt";
