+++ apps/api/src/index.ts
@@ -1,9 +1,4 @@
-import dotenv from "dotenv";
-import path from "path";
-
-dotenv.config();
-dotenv.config({ path: path.resolve(__dirname, "../../.env") });
-dotenv.config({ path: path.resolve(__dirname, "../../../.env") });
+import "dotenv/config";
 
 import { serve } from "@hono/node-server";
 import { createApp } from "./app";
