+++ apps/web/package.json
@@ -12,8 +12,6 @@
   },
   "dependencies": {
     "@costwise/api-client": "workspace:*",
-    "@costwise/db": "workspace:*",
-    "@costwise/domain": "workspace:*",
     "@costwise/shared": "workspace:*",
     "@fontsource/jetbrains-mono": "^5.2.8",
     "@hookform/resolvers": "^5.1.1",
@@ -21,16 +19,12 @@
     "@radix-ui/react-select": "^2.3.0",
     "@radix-ui/react-slot": "^1.2.5",
     "@vercel/analytics": "^1.5.0",
-    "@vercel/blob": "^1.1.1",
-    "bcrypt": "^6.0.0",
     "better-auth": "^1.7.1",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "dotenv": "^17.2.0",
-    "drizzle-orm": "^0.44.5",
     "lucide-react": "^0.525.0",
     "next": "15.3.8",
-    "pg": "^8.16.3",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.60.0",
@@ -46,13 +40,11 @@
     "@testing-library/dom": "^10.4.1",
     "@testing-library/jest-dom": "^6.8.0",
     "@testing-library/react": "^16.3.0",
-    "@types/bcrypt": "^6.0.0",
     "@types/jest": "^30.0.0",
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
     "@types/uuid": "^10.0.0",
-    "drizzle-kit": "^0.31.4",
     "eslint": "^9",
     "eslint-config-next": "15.3.4",
     "jest": "^30.1.3",
