+++ packages/db/src/db.ts
@@ -1,10 +1,3 @@
-import dotenv from 'dotenv';
-import path from 'path';
-
-dotenv.config();
-dotenv.config({ path: path.resolve(__dirname, '../../.env') });
-dotenv.config({ path: path.resolve(__dirname, '../../../.env') });
-
 import { drizzle } from 'drizzle-orm/node-postgres';
 import { Pool } from 'pg';
 import * as schema from './schema';
