+++ packages/shared/src/specialTypes.ts
@@ -1,4 +1,4 @@
-import { Supplier } from "./recipe";
+import { Recipe, Supplier } from "./recipe";
 
 export type RecipeIngredientFromDB = {
   id: number;
@@ -165,3 +165,37 @@ export type Metadata = {
   itemsPerPage: number;
   offset: number;
 };
+
+export interface RecipeAnalytics {
+  avgProfitMargin: string | null;
+  avgFoodCost: string | null;
+  totalRecipes: number;
+}
+
+export interface CategoryAnalytics {
+  category: "starter" | "main" | "dessert";
+  count: number;
+  avgFoodCost: string | null;
+}
+
+export interface MarginHighlights {
+  topPerformers: Recipe[];
+  attentionNeeded: Recipe[];
+}
+
+export interface HighImpactIngredient {
+  id: string;
+  name: string;
+  icon: string | null;
+  usage: number;
+  category: string;
+}
+
+export interface IngredientAnalytics {
+  totalIngredients: number;
+}
+
+export interface OperationResult {
+  id: string;
+}
+
